パリイの振りを攻撃判定のある強化攻撃に変更し、少し前進するように変更。ボスのフィボナッチ数列軌道回転切りがバグっていたので、速度を現在の前進方向に回転させるように変更。フィボナッチ数列構造体の、sinとcosが逆だったので修正。一部warningが出ていた部分を修正。 デモ中にHPゲージの更新処理をして例外になってしまうのを修正。

+++ アクションゲーム/Camera.h
@@ -47,7 +47,7 @@ public:
 
 	void ResetTarget() { m_TargetObject = nullptr; } // 注視点をリセット
 
-	void StartVibration(float amplitude, float frequency, float duration = -1.0f) {
+	void StartVibration(float amplitude, float frequency, int duration = -1) {
 		if (cameraInputFg) m_vib.Start(amplitude, frequency, duration); // 操作有効時のみ
 	} // カメラ振動開始
